Move Conditional into the Circuit component.
